move stuff around and split up tests
